Move common methods to abstract VerbHandler class
